upload project show traning

+++ app/Http/Controllers/Admin/ProductController.php
@@ -5,14 +5,29 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Image;
 use App\Models\product;
 use App\Models\Category;
-use Illuminate\Http\Request;
+use App\Models\Language;
 
+use App\Models\Supplier;
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
 use function Pest\Laravel\delete;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductRequest;
+use App\Models\Attribute;
+use App\Models\AttributeValue;
+use App\Models\ProductAttributeValue;
+use Exception;
+use Illuminate\Support\Facades\DB;
 
 class ProductController extends Controller
 {
+    public $locales;
+    public function __construct()
+    {
+        $this->locales = Language::active()->get()->pluck('code')->toArray();
+        // dd($this->locales);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -28,21 +43,42 @@ class ProductController extends Controller
     public function create()
     {
         $categories = Category::active()->get();
-        return view('admins.products.create', compact('categories'));
+        $suppliers = Supplier::active()->select('id', 'name')->get();
+        $attributes = Attribute::active()->get();
+        return view('admins.products.create', compact('categories', 'suppliers', 'attributes'));
     }
 
     /**
      * Store a newly created resource in storage.
      */
+
+    private function findOrCreateAttributeValue($attributeId, $value, $locale)
+    {
+        $attributeValue = AttributeValue::whereTranslation('value', $value, $locale)->first();
+        if (!$attributeValue) {
+            $attributeValue = new AttributeValue();
+            $attributeValue->translateOrNew($locale)->value = $value;
+            $attributeValue->attribute_id = $attributeId;
+            $attributeValue->save();
+        }
+
+        return $attributeValue;
+    }
     public function store(ProductRequest $request)
     {
-        // dd($request->all());
-        // dd($request->image);
+
+
+
         $product = new product();
-        $product->name = $request->name;
+
+        foreach ($this->locales as $locale) {
+            $product->translateOrNew($locale)->name = $request->get('name_' . $locale);
+            $product->translateOrNew($locale)->description = $request->get('description_' . $locale);
+        }
         $product->price = $request->price;
         $product->quantity = $request->quantity;
         $product->category_id = $request->category_id;
+        $product->supplier_id = $request->supplier_id;
         $product->save();
 
         $product->image()->create([
@@ -56,6 +92,26 @@ class ProductController extends Controller
                     'path' => uploadImage($gallery, 'products'),
                     'type' => 'gallery'
                 ]);
+            }
+        }
+
+        foreach ($request->attributeValue as $attributeName => $item) {
+            // dd($request->$attributeName);
+            if ($request->$attributeName) {
+                $attribute = Attribute::findOrFail($request->$attributeName);
+                foreach ($this->locales as $locale) {
+                    foreach ($item[$locale] as $value) {
+                        try {
+                            DB::beginTransaction();
+                            $attributeValue = $this->findOrCreateAttributeValue($attribute->id, $value, $locale);
+                            $product->attributeValues()->attach($attributeValue->id);
+                            DB::commit();
+                        } catch (Exception $e) {
+                            DB::rollBack();
+                            throw new Exception($e->getMessage());
+                        }
+                    }
+                }
             }
         }
         flash()->success('Product Created Successfully');
@@ -77,7 +133,39 @@ class ProductController extends Controller
     public function edit(Product $product)
     {
         $categories = Category::active()->get();
-        return view('admins.products.edit', compact('product', 'categories'));
+        $suppliers = Supplier::active()->select('id', 'name')->get();
+        $attributes = Attribute::active()->get();
+
+
+        $attributesProduct = [];
+        foreach ($this->locales as $locale) {
+            foreach ($attributes as $key => $attribute) {
+                $values = $product->attributeValues->filter(function ($val) use ($attribute) {;
+                    return $val->attribute_id == $attribute->id;
+                });
+                $valueWithTrans = [];
+                foreach ($values as $one) {
+                    if (!empty($one->translate($locale)->value)) {
+                        $valueWithTrans[] = $one->translate($locale)->value;
+                    }
+                }
+                if (!empty($values)) {
+                    $attributesProduct[$attribute->name][$locale] = $valueWithTrans;
+                }
+            }
+        }
+
+        // foreach ($attributesProduct as $key => $translations) {
+        //     for ($i = 0; $i < count($translations['ar']); $i++) {
+
+        //         dd($translations['en'][$i] ?? '');
+        //     }
+        // }
+
+        // unset($attributesProduct['Color']);
+
+        // dd($product->attributeValues[0]->attribute);
+        return view('admins.products.edit', compact('product', 'categories', 'suppliers', 'attributes', 'attributesProduct'));
     }
 
     /**
@@ -86,16 +174,23 @@ class ProductController extends Controller
     public function update(ProductRequest $request, Product $product)
     {
 
-        // dd($request->all());
-        // dd($product->image->path);
-        $product->name = $request->name;
+        // dd($request->attributeValue);
+
+
+        foreach ($this->locales as $locale) {
+            $product->translateOrNew($locale)->name = $request->get('name_' . $locale);
+            $product->translateOrNew($locale)->description = $request->get('description_' . $locale);
+        }
         $product->price = $request->price;
         $product->quantity = $request->quantity;
         $product->category_id = $request->category_id;
+        $product->supplier_id = $request->supplier_id;
 
         if ($request->hasFile('image')) {
-            deleteImage($product->image->path, 'products');
-            $product->image->delete();
+            if ($product->image) {
+                deleteImage($product->image->path, 'products');
+                $product->image->delete();
+            }
             $product->image()->create([
                 'path' => uploadImage($request->file('image'), 'products'),
                 'type' => 'main'
@@ -103,20 +198,40 @@ class ProductController extends Controller
         }
 
         if ($request->hasFile('galleries')) {
-            foreach ($request->file('galleries') as $one) {
+            foreach ($request->file('galleries') as $image) {
                 $product->gallery()->create([
-                    'path' => uploadImage($one, 'products'),
+                    'path' => uploadImage($image, 'products'),
                     'type' => 'gallery'
                 ]);
             }
         }
 
+
+        $product->attributeValues()->detach();
+        foreach ($request->attributeValue as $attributeName => $item) {
+            if ($request->$attributeName) {
+                $attribute = Attribute::findOrFail($request->$attributeName);
+                foreach ($this->locales as $locale) {
+                    foreach ($item[$locale] as $value) {
+                        try {
+                            DB::beginTransaction();
+                            $attributeValue = $this->findOrCreateAttributeValue($attribute->id, $value, $locale);
+                            // dd($attributeValue->value);
+                            // problem here
+                            $product->attributeValues()->attach($attributeValue->id);
+                            DB::commit();
+                        } catch (Exception $e) {
+                            DB::rollBack();
+                            throw new Exception($e->getMessage());
+                        }
+                    }
+                }
+            }
+        }
+
         $product->save();
 
-
-
         flash()->success('Product Updated Successfully');
-
         return redirect()->route('admin.product.index');
     }
 
@@ -125,7 +240,10 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        deleteImage($product->image()->path, 'products');
+        try {
+            deleteImage($product->image->path, 'products');
+        } catch (\Throwable $th) {
+        }
         foreach ($product->gallery() as $one) {
 
             deleteImage($one->path, 'products');
